profile for user, edit security config; add function for add station from googlemap and REST controller for it,
add jsp for user profile, change href for navbar, add meth. to main controller;
add google map page, rest controller, js file, methods (service, dao, js layers) that add station with name and coordinates in DB.
change latitude, longitude type from long to double,

+++ src/main/java/schedule/entity/StationEntity.java
@@ -14,10 +14,10 @@ public class StationEntity {
     private String stationName;
 
     @Column(name = "latitude")
-    private long latitude;
+    private double latitude;
 
     @Column(name = "longitude")
-    private long longitude;
+    private double longitude;
 
     public int getId() {
         return id;
@@ -35,19 +35,19 @@ public class StationEntity {
         this.stationName = stationName;
     }
 
-    public long getLatitude() {
+    public double getLatitude() {
         return latitude;
     }
 
-    public void setLatitude(long latitude) {
+    public void setLatitude(double latitude) {
         this.latitude = latitude;
     }
 
-    public long getLongitude() {
+    public double getLongitude() {
         return longitude;
     }
 
-    public void setLongitude(long longitude) {
+    public void setLongitude(double longitude) {
         this.longitude = longitude;
     }
 }
